Fix adding new categories from edit product. Resolves trello #1347.
Previously, advanced metaboxes were copied to Exchange's tabs by
copying the html in javascript, and pasting it in the new elements
position. The majority of WordPress's hooks operate on load, and are
binded directly to the element, not on the document. This meant that
things like the add categories hooks had to be re-registered by
Exchange.

Instead of copying the HTML, detach and append the elements so that all
of the original events are preserved. As a side-benefit we get to use
the fancy editor shortcuts.

Needs cross-browser testing.