Begin making coupons more OO for easier testing and extending Coupon types are now registered with an additional class parameter. To maintain BC, a default class of IT_Exchange_Coupon is provided. But coupon types should register their own class.
A new class is introduced IT_Exchange_Cart_Coupon for the basic coupons
core-addon. All of the 'additional parameters' have been marked as
properties and are deprecated. They have corresponding getters to make
the coupon logic more obvious to 3rd-party developers and to allow for
proper mocking.

Some minor changes have been to other API methods, but BC is maintained
This is vert much a first pass. New methods are not being used
internally, yet...