Add methods for manipulating coupon uses Introduces IT_Exchange_Coupon::use_coupon() to be called whenever a coupon is processed on a transaction. That method in turn calls, IT_Exchange_Coupon::increment_usage(). This is overwritten in coupon type child classes. An additional method, IT_Exchange_Coupon::decrement_usage() will be called when a transaction is cancelled or fails.
Cart coupons adds three additional methods to handle these scenarios,
modify_quantity_available(), bump_customer_coupon_frequency() and
reduce_customer_coupon_frequency(). The latter two functions accept
a transaction parameter to associate a usage with a transaction. This
allows for the usage to be reversed. A bug is also fixed with Guest
Checkout and Transient Transactions.

As a result, the following functions are deprecated:

- it_exchange_basic_coupons_bump_customer_coupon_frequency()
- it_exchange_basic_coupons_modify_coupon_quantity_on_transaction()
- it_exchange_basic_coupons_bump_for_customer_on_checkout()