Use a unique ID when updating a trial payment's method ID
Free trial payments in PayPal don't have an associated txn ID. Before
the free trial is converted to a full subscription, we associate the
trial payment with a transient transaction ID.

Previously, when the first actual payment came through, we set the
trial's method ID to that payment's PayPal transaction ID. This caused
the child transaction to never be created.

To remedy this, we set the method ID to be an md5 of PayPal's txn ID
for the first payment.