Change how coupon dates are displayed
The start date and end date columns are conslidated into a new
availability column. The availible column is renmaed to quantity.
Dates are intelligently displayed depending on if the two dates are
in the same month, or year. Sorting has been removed for that column.
We'll be adding a separate filter option for that shortly.