Store the allotted coupon quantity for tracking Previously, Exchange keeped track of the remaining usages for a particular coupon. This number was decremented after every use. This meant we couldn't tell the original number of usages allotted.
Now, we keep track of the original amount of coupons allotted.
Additionally, the edit coupon interface input for quantity refers to
the allotted coupon quantity. For convenience, the total coupon uses
are listed underneath the quantity input.

The coupons list table also shows the number of usages for each coupon
along with the allotted amount.

This will require an upgrade routine to work fully.