Generate ghost menu item URLs on the fly
Change up how Exchange menu items are stored and displayed to the user.
Previously, we used the 'custom' menu item type to store ghost menu
items. This meant the URL at the time of creation was saved, and
results in methods like 'maybe_update_ghost_pages_in_wp_nav_menus' to
attempt to update those URLs whenever they might need to be changed.
 This also generated the unnecessary custom link UI.

Menu items are now stored in the DB as the 'it-exchange-ghost-page'
item type. The menu item label now reads 'iThemes Exchange' instead of
'Custom Link'. The URL property is now generated on the fly based on
the 'object' – that property holds Exchange's page name.

This should ensure that the menu item URLs are always correct.