Introduce middleware for sending emails
Introduces IT_Exchange_Email_Middleware interface, and
IT_Exchange_Email_Middleware_Handler for working with email middleware.

The tag replacers will soon operate as middleware, as opposed to being
configured in each sender. This will make the system less fragile to
changes, and easier for more modifications to be made, such as auto
linking, and emogrifying.