Add support for storing arbitrary data on a cart.
In most cases, line items should be used for storing related data,
however in some instances, like EU VAT, it becomes necessary to store
global information that isn't specific to a line item. For example an
exemption # which would actually stop the construction of line items.

This "cart metadata" is backed by the repository classes. In most cases
this will be session data while the cart is 'active' and then converted
to post meta after the cart has been purchased. The meta keys are
prefixed with '_it_exchange_cart_' to prevent conflicts.