Only mark the SW as rendered if in a sidebar or loop_start has fired.
Some plugins and/or themes will call `the_content` on non main content
before the page has rendered. Because of
`IT_Exchange_Pages::fallback_filter_for_page_template()` this can have
the side-effect of rendering the SW early on product pages.

Now, we only mark the SW as being rendered if we are currently inside
a sidebar ( to account for sidebars rendering before the main content
area, or when the `loop_start` hook has been fired.

In the future, we should look at reducing the scope of the fallback
filter so it only renders on the correcet content.