Add Cart Decorator and Cart Feedback middleware.
Cart decorator automatically sets the cart on the request object. This allows
us to reuse the cart instance.

By reusing the cart instance we are able to keep track of the cart feedback
that has been added throughout the request. The Cart Feedback middleware
renders this feedback in the response on non-GET requests.