Add support for specifying a 'redirect_to' URL.
This allows for customizing where the customer will be redirected
to after purchase. Most handy for things like PayPal purchases initiated
through the API.

The redirect URL is sanitized and validated by core, so by default a user
can only be redirected to a URL on the website. This can be modified
using a filter.