First pass at sharing address records.
By default, one address record will exist for each unique combination of
address, customer, and type. Each transaction will point to that shared
value. If an update to a transaction's address is made through the
repository, and that address is shared amongst multiple other
transactions, it will be split off into a new record so as not to
improperly, update other transactions.