Work around PayPal not always sending txn_id if possible.
On reattempts, PayPal omits the txn_id field for an unknown reason. A previous commit
made it so we don't rely on a reattempt for normal behavior. However, if a fatal error
is encountered during the processing, we will need to use the reattempt IPN. Try to
work around it as best as possible by using the 'custom' field.