Freeze order total after cart is marked as purchase.
PP standard and secure verify that the total coming from PayPal is the same
as the total of the cart. Freeze the cart total right before the customer is
sent to PayPal so we can verify the totals more accurately. Otherwise, it could
be possible for the calculated total to be different if, for example, a tax rate
changes during a longer IPN clearing time.