Pre-emptively fetch customer w/ embeds during backbone checkout, even if not empty.
Previously we'd check if the customer has been loaded at all, and if not fetch.
However, this means that if the cart is using embeds, the customer will be
preloaded, but without her tokens,. THis would end up requiring an additional
http request once a gateway using tokens is selected.