Introduce AuthScope API for managing permissions in the REST API.
Adds an `AuthScope` interface with implementations for CustomerAuth, GuestAuth,
and PublicAuth. This allows us to generalize the user_can_() permission callbacks
to not have to worry about whether the user exists or not, or the user is a guest
or not.

`AuthScope` defines a single method `can()` that accepts a capability and a list
of arguments. This is meant to mirror `current_user_can()`. Each Scope implements
its own mechanism of checking whether the object can perform the given capability.
CustomerAuth delegates to `user_can()` while the other two are more specialized.

The eventual goal here, in perhaps v2.1, so allow for the administrator to issue
authentication keys that have a limited set of capabilities while still
representing that user. For example an key pair that only has support for viewing
transactions, even though the underlying user has permission to edit transactions.

Additionally, Guest Authentication now uses its own Authorization scheme instead
of repurposing basic auth. A client should now send an Authorization header like:
Authorization: ITHEMES-EXCHANGE-GUEST email="guest@example.org" to authenticate
as a guest shopper.