Introduce PAI functions for determining if a product can be purchased by a guest customer. Add guest checkout tests.
Introduces the following API functions.
- it_exchange_can_line_item_be_purchased_by_guest()
- it_exchange_can_cart_be_purchased_by_guest()
- it_exchange_is_guest_checkout_enabled()

Filters are also in place for customizing the functionality. For example, forbiding membership
products to be purchased as a guest customer. Previously, this was only restricted in the UI.
This adds the restriction to the backend so it can't be bypassed by calling the SW API
directly or through the REST API.

Additionally, make error messages descriptive for the Guest customer REST API authentication.