Introduce Scopable line items.
The line item API revolves around the idea of a line item being split up
amongst multiple line items as aggregatables. For example, a coupon
line item is added to each product in the cart so that accurate totals
are provided. This, however, has the side effect of saving a lot of
duplicate information. For example, each coupon line item saves info
about the coupon code, the discount amount, discount type, etc...

The `ITE_Scopable_Line_Item` interface provides a way to access the
"parent" line item for a scoped line item. For example, the global coupon
line item is accessible from the per-product line item. Additionally,
the interface defines a function that specifies the shared parameters
between scoped line items.

A line item that `::is_scoped()` proxies `*_param()` functions to the
line item the item is scoped from for any of shared parameters. This
allows for consuming code to access those parameters transparently,
but for repositories to reduce the storage needs. This is leveraged in
the transaction repository as each line item param is stored as metadata,
which does not currently allow bulk insert.

This is currently only implemented for coupon line items. Will be added to
shipping and taxes shortly.