First step in making sessions require a commit.
Currently, every time any change is made to the session, it is immediately
persisted to the database. When adding items to the cart, this can end up
making multiple writes.

This introduces an ITE_Committable_Session interface and it_exchange_commit_session()
api function. The session is committed on the shutdown hook and after the
Super Widget processing is completed.

For backwards compatibility, the default session handler still commits after
every change. Instead a _doing_it_wrong() notice is issued if a change is made
to the session without it being manually committed. In 2.1 or 2.2 we'll make
the change to require the session to be committed.